Update contract clients
