+++ apps/scoutgame/components/common/NFTPurchaseForm/NFTPurchaseForm.tsx
@@ -5,7 +5,7 @@ import { log } from '@charmverse/core/log';
 import { ActionType, ChainId, SwapDirection } from '@decent.xyz/box-common';
 import { BoxHooksContextProvider, useBoxAction } from '@decent.xyz/box-hooks';
 import { Box, Button, Stack, ToggleButton, ToggleButtonGroup, Typography } from '@mui/material';
-import { BuilderNFTSeasonOneClient } from '@packages/scoutgame/builderNfts/builderNFTSeasonOneClient';
+import { builderContractReadonlyApiClient } from '@packages/scoutgame/builderNfts/clients/builderContractReadClient';
 import {
   builderContractAddress,
   builderNftChain,
@@ -36,12 +36,6 @@ import { NumberInputField } from '../Fields/NumberField';
 
 import type { ChainOption } from './ChainSelector';
 import { BlockchainSelect, getChainOptions } from './ChainSelector';
-
-const readonlyApiClient = new BuilderNFTSeasonOneClient({
-  chain: builderNftChain,
-  contractAddress: builderContractAddress,
-  publicClient: getPublicClient(builderNftChain.id)
-});
 
 export type NFTPurchaseProps = {
   builder: MinimalUserInfo & { price?: bigint; nftImageUrl?: string | null };
@@ -93,7 +87,7 @@ function NFTPurchaseButton({ builder }: NFTPurchaseProps) {
   });
 
   useEffect(() => {
-    readonlyApiClient.getProceedsReceiver().then(setTreasuryAddress);
+    builderContractReadonlyApiClient.getProceedsReceiver().then(setTreasuryAddress);
   }, []);
 
   const refreshBalance = useCallback(async () => {
@@ -141,7 +135,7 @@ function NFTPurchaseButton({ builder }: NFTPurchaseProps) {
 
   const refreshAsk = useCallback(
     async ({ _builderTokenId, amount }: { _builderTokenId: bigint | number; amount: bigint | number }) => {
-      const _price = await readonlyApiClient.getTokenPurchasePrice({
+      const _price = await builderContractReadonlyApiClient.getTokenPurchasePrice({
         args: { amount: BigInt(amount), tokenId: BigInt(_builderTokenId) }
       });
       setPurchaseCost(_price);
@@ -153,7 +147,7 @@ function NFTPurchaseButton({ builder }: NFTPurchaseProps) {
     setFetchError(null);
     try {
       setIsFetchingPrice(true);
-      const _builderTokenId = await readonlyApiClient.getTokenIdForBuilder({ args: { builderId } });
+      const _builderTokenId = await builderContractReadonlyApiClient.getTokenIdForBuilder({ args: { builderId } });
 
       setBuilderTokenId(_builderTokenId);
