+++ packages/scoutgame/src/builderNfts/clients/builderContractReadClient.ts
@@ -1,0 +1,11 @@
+import { getPublicClient } from '@packages/onchain/getPublicClient';
+
+import { builderContractAddress, builderNftChain } from '../constants';
+
+import { BuilderNFTSeasonOneImplementation01Client } from './builderNFTSeasonOneClient';
+
+export const builderContractReadonlyApiClient = new BuilderNFTSeasonOneImplementation01Client({
+  chain: builderNftChain,
+  contractAddress: builderContractAddress,
+  publicClient: getPublicClient(builderNftChain.id)
+});
